requirements.txt
accelerate==0.17.1  # Load 8bit
bitsandbytes==0.37.2  # Load 8bit
datasets==2.12.0
einops==0.6.1
fire==0.5.0
pillow==9.5.0
pydantic==1.10.6
torch==2.0.0
torchmetrics==0.11.4
torchvision==0.15.1
tqdm==4.65.0
transformers==4.28.1
